backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
python-dotenv
openai
streamlit
requests
